Fix edit on empty tag
When adding a lap to the daily records, if the tag were empty, then the
html p element wouldn't contain any height nor width to be clickable to
edit. This commit adds a minimum widht and height the element has to
contain even if empty.

In the scenario the tag overflow the space for its entries, it also
displays ellipsis.