Fix overlapping records in cumulatice chart
When 2 records showed overlapping periods, the chart would create a
zig-zag pattern which is not intended. In the context of the cumulative
chart, if 2 records overlap, the behavior defined in this commit are
that the records merge together creating a single line.

The cumulative code was also refactored to seperate some processing into
its own lib file.