Add list of tags to store
To provide a list of tags when modifying a tag entry, this commit adds a
list of tags to the global store. When adding a new tag to the list of
tags, the slice also handles updating the local storage. An event
listener is also added specifically to the tags key in the localStorage.