Add controls to tag component
As a user is typing a tag on the input field, a list of tags from the
global state is filtered based on the substring. This tag allows a user
to create a new tag in case the one the user is typing does not exist.
It also enables a user to delete a tag from the list of available tags.